feat: workshop v2
