+++ 01-echo/request.http
@@ -1,0 +1,2 @@
+GET http://localhost:7071/api/HttpTrigger?name=test
+Content-Type: application/json
