--- 01-echoFunction/.gitignore
@@ -1,38 +1,0 @@
-# Build output
-target/
-*.class
-
-# Log file
-*.log
-
-# BlueJ files
-*.ctxt
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-*.zip
-*.tar.gz
-*.rar
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-# IDE
-.idea/
-*.iml
-.settings/
-.project
-.classpath
-
-# macOS
-.DS_Store
-
-# Azure Functions
-local.settings.json
-bin/
-obj/
