--- 02-age-and-gender-recognition/upload.html
@@ -1,82 +1,0 @@
-<html>
-    <head>
-        <title>Image Classification Test</title>
-        <script>
-            
-            function executeFunction(imageAsBase64) {                
-
-                let xhr = new XMLHttpRequest();
-
-                endpointURL = document.getElementById("endpointURL").value
-                xhr.open("POST", endpointURL);
-
-                xhr.responseType = 'blob';
-                xhr.setRequestHeader("Accept", "*/*");
-                xhr.setRequestHeader("Content-Type", "application/json");
-
-                xhr.onload = () => {
-                    var loadingIndicator = document.getElementById("loadingIndicator");
-                    loadingIndicator.style.display = "none";
-                    var dd = URL.createObjectURL(xhr.response);
-                    document.getElementById("imgResult").setAttribute('src', dd);
-                }
-
-                let data = `{"img": "` +imageAsBase64 + `"}`;
-                
-                xhr.send(data);       
-            }
-
-            function encodeImageFileAsURL() {  
-                var loadingIndicator = document.getElementById("loadingIndicator");
-                loadingIndicator.style.display = "block";
-
-                var filesSelected = document.getElementById("inputFileToLoad").files;
-                if (filesSelected.length > 0) {
-                var fileToLoad = filesSelected[0];
-
-                var fileReader = new FileReader();
-
-                fileReader.onload = function(fileLoadedEvent) {
-                    var srcData = fileLoadedEvent.target.result; // <--- data: base64
-
-                    var newImage = document.createElement('img');
-                    newImage.src = srcData;
-
-                    document.getElementById("imgTest").innerHTML = newImage.outerHTML;
-                    console.log("Converted Base64 version is " + document.getElementById("imgTest").innerHTML);
-
-                    executeFunction(srcData.replace('data:image/jpeg;base64,', ''));
-                }
-                fileReader.readAsDataURL(fileToLoad);
-                }
-            } 
-        </script>
-    </head>
-
-    <body>
-        <h4>Description</h4>
-        Sample page for testing image classification functions 
-        <ul>
-            <li>Define endpoint for your Azure function</li>
-            <li>Select an image</li>
-            <li>On selecting a new image the image is converted to base64 and sent to the function</li>
-            <li>The reponse is rendered as a result</li>            
-        </ul>
-
-        <i>Pro Tip: Press F12 to open developer tools and have a look on the console and network tab to see what is going on. ;)</i><br><br>
-
-        Define function endpoint <br>
-        <input id="endpointURL" type="text" value="https://<YOUR_FUNCTION_APP>.azurewebsites.net/api/httptrigger" size="200"/> <br><br>
-
-        Select Image <br>
-        <input id="inputFileToLoad" type="file" onchange="encodeImageFileAsURL();"/>
-        
-        <h4>Result</h4>
-        <img id="imgResult" src="" alt=""  width="500px" />
-        <div id="loadingIndicator" hidden="true"><marquee width = "25%"><h3>Waiting for response. Be amazed by one of the fanciest loading indicators. ;)</h3></marquee></div>
-
-        <h4>Input</h4>
-        <div id="imgTest"></div>
-    </body>
-
-</html>
